+++ @navikt/core/react/src/stepper/context.ts
@@ -1,0 +1,11 @@
+import { createContext } from "react";
+
+interface StepperContextProps {
+  activeStep: number;
+  onStepChange: (step: number) => void;
+  lastIndex: number;
+  orientation: "horizontal" | "vertical";
+  interactive: boolean;
+}
+
+export const StepperContext = createContext<StepperContextProps | null>(null);
